# packages/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/calendars": {
      "get": {
        "description": "Only active, validated calendars that have synced at least once. The Google calendar id and feed URL are never exposed; each item carries a safe googleOpenUrl.",
        "operationId": "PublicCalendarController_list_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCalendarListResponseDto"
                }
              }
            }
          }
        },
        "summary": "List public calendars (read-only, synced from their public ICS feed).",
        "tags": [
          "public-calendars"
        ]
      }
    },
    "/v1/calendars/events": {
      "get": {
        "description": "Pass one `calendar` slug per selected calendar. An empty selection returns an empty list (never \"all calendars\"). Each event carries calendarId and calendarSlug. The result is capped server-side; `meta.truncated` reports a cut list rather than presenting it as complete.",
        "operationId": "PublicCalendarController_aggregated_v1",
        "parameters": [
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {}
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {}
          },
          {
            "name": "calendar",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCalendarEventsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Aggregated events across several selected public calendars.",
        "tags": [
          "public-calendars"
        ]
      }
    },
    "/v1/calendars/google-view-url": {
      "get": {
        "description": "Server-constructed HTTPS URL (one src per calendar). This is a shared VIEW only — it never adds anything to a user's personal Google account.",
        "operationId": "PublicCalendarController_googleViewUrl_v1",
        "parameters": [
          {
            "name": "calendar",
            "required": true,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleViewUrlResponseDto"
                }
              }
            }
          }
        },
        "summary": "A combined calendar.google.com embed URL for the selected calendars.",
        "tags": [
          "public-calendars"
        ]
      }
    },
    "/v1/calendars/{slug}/events": {
      "get": {
        "description": "The result is capped server-side; `meta.truncated` reports a cut list rather than presenting it as complete.",
        "operationId": "PublicCalendarController_single_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "de",
                "en"
              ],
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {}
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "YYYY-MM-DD",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCalendarEventsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Events of a single public calendar in a date range.",
        "tags": [
          "public-calendars"
        ]
      }
    },
...
  },
  "info": {
    "title": "Campus Köthen API",
    "description": "Public read-only API of the Campus Köthen app.\n\nCampus Köthen is an independent, unofficial campus app. It is neither developed\nnor operated by Hochschule Anhalt, nor is it officially endorsed by the university.\n\nAll content endpoints accept `locale=de|en` (default `de`) and report\n`requestedLocale`, `resolvedLocale` and `translationFallback` in their metadata.\nCanteen dish text comes from a German-only source and is never machine-translated.",
    "version": "0.1.0",
    "contact": {
      "name": "Campus Köthen App",
      "url": "https://github.com/Leviora-Studio/campus-koethen",
      "email": ""
    },
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
      "ResponseMetaDto": {
        "type": "object",
        "properties": {
          "requestedLocale": {
            "type": "string",
            "enum": [
              "de",
              "en"
            ],
            "description": "Locale the client asked for."
          },
          "resolvedLocale": {
            "type": "string",
            "enum": [
              "de",
              "en"
            ],
            "description": "Locale actually served."
          },
          "translationFallback": {
            "type": "boolean",
            "description": "True when at least one delivered field came from the German fallback although another locale was requested."
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          },
          "droppedBlockTypes": {
            "description": "Content block types removed because the MVP client does not render them.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastSuccessfulSyncAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last successful canteen synchronisation. Null means: never succeeded."
          },
          "dataStale": {
            "type": "boolean",
            "description": "True when the data is older than the configured threshold."
          },
          "from": {
            "type": "string",
            "description": "Start of the resolved date range (YYYY-MM-DD)."
          },
          "to": {
            "type": "string",
            "description": "End of the resolved date range (YYYY-MM-DD)."
          },
          "truncated": {
            "type": "boolean",
            "description": "True when the server hit its own result ceiling and deliberately cut the list short. The client must say so rather than present the result as complete."
          },
          "featureEnabled": {
            "type": "boolean",
            "description": "False when the timetable integration is switched off server-side."
          },
          "dataState": {
            "type": "string",
            "enum": [
              "ready",
              "pending",
              "unavailable"
            ],
            "description": "ready = data present; pending = enabled but not yet synchronised; unavailable = disabled or permanently without data."
          },
          "timezone": {
            "type": "string",
            "description": "Business timezone the wall-clock times belong to.",
            "example": "Europe/Berlin"
          }
        },
        "required": [
          "requestedLocale",
          "resolvedLocale",
          "translationFallback"
        ]
      },
...
      "PublicCalendarDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "channelSlug": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "colorHex": {
            "type": "string",
            "example": "#5B3FD0"
          },
          "sortOrder": {
            "type": "number"
          },
          "defaultSubscribed": {
            "type": "boolean"
          },
          "dataState": {
            "type": "string",
            "description": "ready | stale",
            "example": "ready"
          },
          "lastSuccessfulSyncAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "dataStale": {
            "type": "boolean"
          },
          "googleOpenUrl": {
            "type": "string",
            "description": "Safe HTTPS link to open this calendar in Google Calendar."
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "colorHex",
          "sortOrder",
          "defaultSubscribed",
          "dataState",
          "dataStale",
          "googleOpenUrl"
        ]
      },
      "PublicCalendarListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicCalendarDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "PublicCalendarEventDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Stable, non-invertible eventKey = base64url(sha256(calendarSlug + \"\\0\" + occurrenceKey))[0..21]."
          },
          "calendarId": {
            "type": "string"
          },
          "calendarSlug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "allDay": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "description": "confirmed | tentative | cancelled"
          }
        },
        "required": [
          "id",
          "calendarId",
          "calendarSlug",
          "title",
          "start",
          "end",
          "allDay",
          "status"
        ]
      },
      "PublicCalendarEventsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicCalendarEventDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "GoogleViewUrlDataDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "A calendar.google.com embed URL for the selected calendars."
          }
        },
        "required": [
          "url"
        ]
      },
      "GoogleViewUrlResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/GoogleViewUrlDataDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}